internal: move server config to internal/serverconfig
