+++ internal/config/config.go
@@ -41,29 +41,6 @@ type AppURL struct {
 	AutoHostname *bool `hcl:"auto_hostname,optional"`
 }
 
-// Server configures the remote server.
-type Server struct {
-	Address string `hcl:"address,attr"`
-
-	// Tls, if true, will connect to the server with TLS. If TlsSkipVerify
-	// is true, the certificate presented by the server will not be validated.
-	Tls           bool `hcl:"tls,optional"`
-	TlsSkipVerify bool `hcl:"tls_skip_verify,optional"`
-
-	// AddressInternal is a temporary config to work with local deployments
-	// on platforms such as Docker for Mac. We need to discuss a more
-	// long term approach to this.
-	AddressInternal string `hcl:"address_internal,optional"`
-
-	// Indicates that we need to present a token to connect to this server.
-	RequireAuth bool `hcl:"require_auth,optional"`
-
-	// AuthToken is the token to use to authenticate to the server.
-	// Note this will be stored plaintext on disk. You can also use the
-	// WAYPOINT_SERVER_TOKEN env var.
-	AuthToken string `hcl:"auth_token,optional"`
-}
-
 // Runner is the configuration for supporting runners in this project.
 type Runner struct {
 	// Enabled is whether or not runners are enabled. If this is false
